+++ src/index.js
@@ -1,35 +1,52 @@
 // @flow
 
-type getInternalFunction<T> = (object: Object) => T
+/**
+ * Returns an object which contains private data of given object.
+ * @type    {Function}
+ * @name    getInternalFn
+ * @param   {Object} object - Object to get private data of.
+ * @return  {Object}        - Object which contains private data of given object.
+ */
+type getInternalFn<T, I> = (object: T) => I
 
 /**
- * Returns a function to get private data.
- * @returns {Function} - Function to get private properties of given object.
+ * Class which provides private data storage.
+ * @hideconstructor
  */
-function internalDataFunction<T>() : getInternalFunction<T> {
-  const internalData = new WeakMap();
+class InternalData<T, I> {
+  data: WeakMap<T, I>;
 
-  if (arguments.length > 0) {
-    throw new Error("This function must be called without arguments. Maybe you forget () in require. require('internal-data')()");
+  /**
+   * Constructor
+   */
+  constructor() {
+    this.data = new WeakMap();
   }
 
   /**
-   * Returns an object which contains private data of given object.
-   * @param   {Object} object - Object to get private data of.
-   * @returns {Object}        - Object which contains private data of given object.
+   * Returns a function to access priavte data object.
+   * @return  {getInternalFn} - Function to get private properties of given object.
    */
-  function getInternal(object: Object) : T {
-    let internal = internalData.get(object);
-
-    if (!internal) {
-      internal = {};
-      internalData.set(object, internal);
-    }
-
-    return ((internal : any) : T);
+  static getFunction(): getInternalFn<T, I> {
+    const internalData: InternalData<T, I> = new InternalData();
+    return object => internalData.get(object);
   }
 
-  return getInternal;
+  /**
+   * Returns private data object for given object.
+   * @param   {Object} object Object to get private data for.
+   * @return  {Object}        Private data of given object.
+   */
+  get(object: T): I {
+    let internal = this.data.get(object);
+
+    if (!internal) {
+      internal = (({}: any): I);
+      this.data.set(object, internal);
+    }
+    return internal;
+  }
 }
 
-module.exports = internalDataFunction;
+
+export default InternalData;