+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### 0.0.0 - 2016-03-23
+### 0.1.0 - 2017-11-24
 
-#### Added/Changed etc.
+#### Added
 
-* Details
+* Object oriented and function interfaces.
+
+### Removed
+
+* Require and execute interface.